♻️ [Refactoring]: #267 src/commands/args.rs を用途別サブモジュールに分割
- 既存5型 (ListOutputArgs / Single|MultiTargetArgs / Interactive|SyncScopeArgs) を args/{output,target,scope}.rs に移設
- args.rs は mod 宣言と pub use のみに縮退し、旧パス crate::commands::args::* を維持
- doc コメントと help 文字列は一字一句保ち、--help 出力に差分を出さない
